refactor: extract pivot renderer customization from form tabs
