+++ src/components/Form/components/ui/tabs/Tabs.tsx
@@ -1,7 +1,5 @@
 import * as React from "react";
 import { TabComponents, type ITabsComponents } from "./components";
-import { getPivotStyles } from "./styles";
-import { useTheme } from "@fluentui/react";
 
 export type TabLikeChild = React.ReactElement<{ id: string; label?: string; }>;
 
@@ -16,18 +14,10 @@ export interface IFormTabsProps {
 export const Tabs = (props: IFormTabsProps) => {
     const { expandedTab, children, onTabChange } = props;
     const components = { ...TabComponents, ...props.components };
-    const theme = useTheme();
-    const styles = getPivotStyles(theme);
 
     return components.onRenderTabs({
-        overflowBehavior: 'menu',
-        selectedKey: expandedTab,
-        className: styles.pivotContainer,
-        styles: {
-            root: styles.pivot,
-            itemContainer: styles.itemContainer
-        },
-        children: children,
-        onLinkClick: (item) => onTabChange?.(item?.props.itemKey!),
+        expandedTab,
+        children,
+        onTabChange,
     })
 };
