+++ src/components/Form/components/ui/tabs/components/pivot/Pivot.tsx
@@ -1,18 +1,30 @@
-import { Pivot as FluentPivot, PivotItem, useTheme } from "@fluentui/react";
+import { PivotItem, useTheme } from "@fluentui/react";
 import { ITabsComponentProps } from "../components";
 import React, { useMemo } from "react";
 import type { TabLikeChild } from "@components/Form/components/ui/tabs/Tabs";
-import { getPivotItemStyles } from "./styles";
+import { getPivotStyles } from "./styles";
+import { IPivotComponents, PivotComponents } from "./components";
 
+interface IPivotProps extends ITabsComponentProps {
+    components?: Partial<IPivotComponents>;
+}
 
-export const Pivot = (props: ITabsComponentProps) => {
+export const Pivot = (props: IPivotProps) => {
+    
+    const components = { ...PivotComponents, ...props.components };
     const childrenArray: TabLikeChild[] = React.Children.toArray(props.children).filter(child => React.isValidElement(child)) as TabLikeChild[];
-    const styles = useMemo(() => getPivotItemStyles(), []);
+    const theme = useTheme();
+    const styles = useMemo(() => getPivotStyles(theme), [theme]);
 
-    return <FluentPivot
-        {...props}
-        >
-        {childrenArray.map(child => {
+    return components.onRenderPivot({
+        overflowBehavior: 'menu',
+        className: styles.pivotContainer,
+        styles: {
+            root: styles.pivot,
+            itemContainer: styles.itemContainer
+        },
+        onLinkClick: (item) => props.onTabChange?.(item?.props.itemKey!),
+        children: childrenArray.map(child => {
             if (!child.props.id) throw new Error("Tab child is missing required 'id' prop");
             return <PivotItem
                 key={child.props.id}
@@ -22,7 +34,6 @@ export const Pivot = (props: ITabsComponentProps) => {
                 headerButtonProps={{ "data-id": `tab-header-${child.props.id}` }}
                 children={child}
             />;
-        })}
-
-    </FluentPivot>;
+        })
+    })
 };
