+++ src/components/Form/components/ui/tabs/components/pivot/components.tsx
@@ -1,0 +1,9 @@
+import { IPivotProps, Pivot } from "@fluentui/react";
+
+export interface IPivotComponents {
+    onRenderPivot: (props: IPivotProps) => JSX.Element;
+}
+
+export const PivotComponents: IPivotComponents = {
+    onRenderPivot: (props) => <Pivot {...props} />
+}
