+++ src/components/Form/components/ui/tabs/components/pivot/index.ts
@@ -1,1 +1,2 @@
 export * from './Pivot';
+export * from './components';
