+++ src/components/Form/components/ui/tabs/components/pivot/styles.ts
@@ -1,4 +1,4 @@
-import { mergeStyleSets } from "@fluentui/react";
+import { ITheme, mergeStyleSets } from "@fluentui/react";
 
 const FLEX_STYLES = {
     display: 'flex',
@@ -6,8 +6,19 @@ const FLEX_STYLES = {
     flexGrow: 1,
 };
 
-export const getPivotItemStyles = () => {
+export const getPivotStyles = (theme: ITheme) => {
     return mergeStyleSets({
+        pivot: {
+            //boxShadow: theme.effects.elevation4,
+            height: 48,
+            //zIndex: 1
+        },
+        pivotContainer: {
+            ...FLEX_STYLES,
+        },
+        itemContainer: {
+            ...FLEX_STYLES
+        },
         pivotItem: {
             ...FLEX_STYLES
         }
